EICNET-2092: Rename access check to GroupViewPageAccessChecker and make sure it checks access only for view pages.
