+++ lib/modules/eic_groups/eic_groups.services.yml
@@ -51,10 +51,10 @@ services:
     tags:
       - { name: access_check, applies_to: _group_invitation_bulk }
   eic_groups.access_check.group_pages:
-    class: Drupal\eic_groups\Access\GroupPagesAccessChecker
-    arguments: ['@eic_groups.helper']
+    class: Drupal\eic_groups\Access\GroupViewPageAccessChecker
+    arguments: ['@current_route_match']
     tags:
-      - { name: access_check, applies_to: _group_pages_access_check }
+      - { name: access_check, applies_to: _group_view_pages_access_check }
   eic_groups.route_subscriber:
     class: Drupal\eic_groups\Routing\RouteSubscriber
     tags:
